requirements.txt
Flask==3.0.3
twilio==9.2.3
google-generativeai==0.7.2
websockets==11.0.3     
soundfile==0.12.1
numpy==1.26.0
python-dotenv==1.0.0

